+++ backend/src/main/java/io/ndk/backend/service/impl/TaskServiceImpl.java
@@ -1,6 +1,7 @@
 package io.ndk.backend.service.impl;
 
 import io.ndk.backend.Mappers.Mapper;
+import io.ndk.backend.Mappers.impl.TaskMapper;
 import io.ndk.backend.dto.TaskDto;
 import io.ndk.backend.dto.request.TaskRequest;
 import io.ndk.backend.entity.Category;
@@ -25,34 +26,47 @@ public class TaskServiceImpl implements TaskService {
     private final TaskRepository taskRepository;
     private final UserRepository userRepository;
     private final CategoryRepository categoryRepository;
-    private final Mapper<Task, TaskDto> mapper;
+    private final TaskMapper mapper;
 
     @Override
     public List<TaskDto> getTaskByUser(String email) {
-        User user = userRepository.findByEmail(email).orElseThrow(() -> new CustomException(BusinessErrorCodes.NO_SUCH_EMAIL));
+        User user = getCurrentUser(email);
         return taskRepository.findByUser(user).stream().map(mapper::mapTo).collect(Collectors.toList());
     }
 
     @Override
     public List<TaskDto> getTasksWithoutCategory(String email) {
-        User user = userRepository.findByEmail(email).orElseThrow(() -> new CustomException(BusinessErrorCodes.NO_SUCH_EMAIL));
+        User user = getCurrentUser(email);
         return taskRepository.findByUser(user).stream().filter(t -> t.getCategory() == null).map(mapper::mapTo).collect(Collectors.toList());
     }
 
     @Override
-    public List<TaskDto> getTaskByCategoryId(Long categoryId) {
-        Category category = categoryRepository.findById(categoryId).orElseThrow(() -> new CustomException(BusinessErrorCodes.NO_SUCH_CATEGORY));
-        return taskRepository.findByCategory(category).stream().map(mapper::mapTo).collect(Collectors.toList());
+    public List<TaskDto> getTaskByCategoryId(Long categoryId, String email) {
+        User user = getCurrentUser(email);
+        Category category = getCategoryById(categoryId);
+
+        if (!category.getUser().getEmail().equals(user.getEmail())) {
+            throw new CustomException(BusinessErrorCodes.ACCESS_DENIED);
+        }
+
+        return taskRepository.findByCategory(category)
+                .stream()
+                .map(mapper::mapTo)
+                .collect(Collectors.toList());
     }
 
     @Override
     public TaskDto createTask(TaskRequest request, String email) {
-        User user = userRepository.findByEmail(email).orElseThrow(() -> new CustomException(BusinessErrorCodes.NO_SUCH_EMAIL));
+        User user = getCurrentUser(email);
         Category categoryEntity = null;
         if(request.getCategoryId() != null) {
-            categoryEntity = categoryRepository.findById(request.getCategoryId()).orElseThrow(() -> new CustomException(BusinessErrorCodes.NO_SUCH_CATEGORY));
+            categoryEntity = getCategoryById(request.getCategoryId());
+
+            if (!categoryEntity.getUser().getEmail().equals(user.getEmail())) {
+                throw new CustomException(BusinessErrorCodes.ACCESS_DENIED);
+            }
         }
-        Task task = new Task().builder()
+        Task task = Task.builder()
                 .title(request.getTitle())
                 .description(request.getDescription())
                 .category(categoryEntity)
@@ -64,18 +78,34 @@ public class TaskServiceImpl implements TaskService {
     }
 
     @Override
-    public TaskDto getTaskById(Long id) {
-        Task task = taskRepository.findById(id).orElseThrow(() -> new CustomException(BusinessErrorCodes.NO_SUCH_TASK));
+    public TaskDto getTaskById(Long id, String email) {
+        User user = getCurrentUser(email);
+        Task task = getTaskById(id);
+
+        if (!task.getUser().getEmail().equals(user.getEmail())) {
+            throw new CustomException(BusinessErrorCodes.ACCESS_DENIED);
+        }
+
         return mapper.mapTo(task);
     }
 
     @Override
-    public TaskDto updateById(Long id, TaskRequest request) {
-        Category categoryEntity = null;
-        if(request.getCategoryId() != null) {
-            categoryEntity = categoryRepository.findById(request.getCategoryId()).orElseThrow(() -> new CustomException(BusinessErrorCodes.NO_SUCH_CATEGORY));
+    public TaskDto updateById(Long id, TaskRequest request, String email) {
+        User user = getCurrentUser(email);
+        Task task = getTaskById(id);
+
+        if (!task.getUser().getEmail().equals(user.getEmail())) {
+            throw new CustomException(BusinessErrorCodes.ACCESS_DENIED);
         }
-        Task task = taskRepository.findById(id).orElseThrow(() -> new CustomException(BusinessErrorCodes.NO_SUCH_TASK));
+
+        Category categoryEntity = null;
+        if (request.getCategoryId() != null) {
+            categoryEntity = getCategoryById(request.getCategoryId());
+            if (!categoryEntity.getUser().getEmail().equals(user.getEmail())) {
+                throw new CustomException(BusinessErrorCodes.ACCESS_DENIED);
+            }
+        }
+
         task.setTitle(request.getTitle());
         task.setDescription(request.getDescription());
         task.setStatus(request.getStatus());
@@ -85,7 +115,29 @@ public class TaskServiceImpl implements TaskService {
     }
 
     @Override
-    public void deleteById(Long id) {
+    public void deleteById(Long id, String email) {
+        User user = getCurrentUser(email);
+        Task task = getTaskById(id);
+
+        if (!task.getUser().getEmail().equals(user.getEmail())) {
+            throw new CustomException(BusinessErrorCodes.ACCESS_DENIED);
+        }
+
         taskRepository.deleteById(id);
+    }
+
+    private User getCurrentUser(String email) {
+        return userRepository.findByEmail(email)
+                .orElseThrow(() -> new CustomException(BusinessErrorCodes.NO_SUCH_EMAIL));
+    }
+
+    private Category getCategoryById(Long id) {
+        return categoryRepository.findById(id)
+                .orElseThrow(() -> new CustomException(BusinessErrorCodes.NO_SUCH_CATEGORY));
+    }
+
+    private Task getTaskById(Long id) {
+        return taskRepository.findById(id)
+                .orElseThrow(() -> new CustomException(BusinessErrorCodes.NO_SUCH_TASK));
     }
 }
